added tests for description & searchValue props
